Move qlpack-generator.ts to /local-queries
